Add protobuf editions grammar file.
